rename twitter plugin to twitter_plugin_gamesdk
